# specs/003-fastapi-better-auth/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI Better-Auth Backend API
  description: |
    Production-ready API with Better-Auth integration for authentication.
    Authentication is handled by Better-Auth (separate service), this API validates JWTs.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user profile
      description: Returns the authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Management Endpoints
  /api/v1/users:
    get:
      tags: [users]
      summary: List all users (Admin only)
      description: Returns paginated list of all users
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, user]
      responses:
        '200':
          description: Users list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: Get user by ID
      description: Returns user details (Admin can view any, users can view self)
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [users]
      summary: Update user
      description: Update user profile (Admin can update any, users can update self)
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/users/{user_id}/role:
    put:
      tags: [users]
      summary: Update user role (Admin only)
      description: Assign or change a user's role
      operationId: updateUserRole
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/users/{user_id}/ban:
    post:
      tags: [users]
      summary: Ban user (Admin only)
      description: Ban a user with optional reason and expiration
      operationId: banUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanUserRequest'
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [users]
      summary: Unban user (Admin only)
      description: Remove ban from a user
      operationId: unbanUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User unbanned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/users/{user_id}/sessions:
    get:
      tags: [users]
      summary: List user sessions
      description: Get all active sessions for a user
      operationId: listUserSessions
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sessions retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [users]
      summary: Revoke all user sessions
      description: Invalidate all sessions for a user (logout everywhere)
      operationId: revokeAllSessions
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better-Auth
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # User Schemas
    User:
      type: object
      required: [id, email, email_verified, role, created_at, updated_at]
      properties:
        id:
          type: string
          example: "01HX..."
        email:
          type: string
          format: email
          example: "user@example.com"
        email_verified:
          type: boolean
          example: true
        name:
          type: string
          nullable: true
          example: "John Doe"
        image:
          type: string
          nullable: true
          format: uri
        role:
          type: string
          enum: [admin, user]
          example: user
        banned:
          type: boolean
          example: false
        ban_reason:
          type: string
          nullable: true
        ban_expires:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
    UsersListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        image:
          type: string
          format: uri
    RoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, user]
    BanUserRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
        expires_at:
          type: string
          format: date-time
          description: Optional expiration time for ban

    # Session Schemas
    Session:
      type: object
      required: [id, user_id, expires_at, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        expires_at:
          type: string
          format: date-time
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    SessionsListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'

    # Pagination
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Admin access required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: User not found
...